test_georeum roll back
